Fix: Brand resolver 분리
